Program divided into modules, <alghorithm> removed
