+++ homework3/hw3.3/frequencySearch.h
@@ -1,0 +1,3 @@
+#pragma once
+
+int frequencySearch(int *searchingArray, int searchingArrayLength);
