+++ homework3/hw3.3/hw3.3.vcxproj
@@ -116,7 +116,10 @@
   </ItemDefinitionGroup>
   <ItemGroup>
     <ClCompile Include="main.cpp" />
-    <ClCompile Include="quickSort.cpp" />
+    <ClCompile Include="frequencySearch.cpp" />
+  </ItemGroup>
+  <ItemGroup>
+    <ClInclude Include="frequencySearch.h" />
   </ItemGroup>
   <Import Project="$(VCTargetsPath)\Microsoft.Cpp.targets" />
   <ImportGroup Label="ExtensionTargets">
