--- homework3/hw3.3/quickSort.cpp
@@ -1,35 +1,0 @@
-#define _CRT_SECURE_NO_WARNINGS
-#include <algorithm>
-
-int arrayPartition(int *partitionArray, int partLeftBorder, int partRightBorder)
-{
-	const int pivot = partitionArray[partLeftBorder];
-	int position = partLeftBorder + 1;
-
-	for (int i = partLeftBorder + 1; i <= partRightBorder; ++i)
-	{
-		if (partitionArray[i] < pivot)
-		{
-			std::swap(partitionArray[i], partitionArray[position]);
-			++position;
-		}
-	}
-
-	std::swap(partitionArray[partLeftBorder], partitionArray[position - 1]);
-
-	return position - 1;
-}
-
-int quickSort(int *qsSourceArray, int sortLeftBorder, int sortRightBorder)
-{
-	if (sortRightBorder < sortLeftBorder)
-	{
-		return 0;
-	}
-
-	int divisor = arrayPartition(qsSourceArray, sortLeftBorder, sortRightBorder);
-	quickSort(qsSourceArray, sortLeftBorder, divisor - 1);
-	quickSort(qsSourceArray, divisor + 1, sortRightBorder);
-
-	return 0;
-}
